Refactored All components. Can now set multiple Co-Authors.
