+++ components/Cards/CardsPage.vue
@@ -1,16 +1,11 @@
 <template>
     <ClientOnly>
-        <StyledCard />
-        <Interactive/>
-        <SlotsCard/>
+        <Stefan/>
     </ClientOnly>
 </template>
 
 <script setup>
-import StyledCard from '@/Authors/Stefan/components/Cards/StyledCard';
-import Interactive from '@/Authors/Stefan/components/Cards/InteractiveCard';
-import SlotsCard from '@/Authors/Stefan/components/Cards/SlotsCard.vue';
-
+import Stefan from '~/components/Cards/Stefan/StefanCardsPage.vue';
 </script>
 
 <style lang="scss" scoped></style>
