+++ components/Cards/Stefan/StefanSearchCards.ts
@@ -3,12 +3,13 @@ import { useComponentSearchStore } from '~/stores/useComponentSearchStore';
 // Search Query
 import { ref } from "vue";
 
-export function useSearchSelects() {
+export function useSearchCards() {
 
 // Reactive data for cards and search query
 const cards = ref([
-  { id: 0, title: "Styled", author: "Stefan" , coAuthor: "None" },
-  { id: 1, title: "Activity", author: "Stefan" , coAuthor: "None" },
+  { id: 0, title: "Styled" , author: "Stefan" , coAuthor: ["None"] },
+  { id: 1, title: "Interactive" , author: "Stefan" , coAuthor: ["None"] },
+  { id: 2, title: "Slots" , author: "Stefan" , coAuthor: ["None"] },
   // Add more cards as needed
 ]);
 
@@ -19,9 +20,11 @@ const matchesSearch = (card: any) => {
   } else if (useComponentSearchStore().filter === "Author"){
     return card.author.toLowerCase().includes(useComponentSearchStore().searchQuery.toLowerCase());
   } else if (useComponentSearchStore().filter === "Co-Author"){
-    return card.coAuthor.toLowerCase().includes(useComponentSearchStore().searchQuery.toLowerCase());
+    return card.coAuthor.some((coAuthor: string) => 
+      coAuthor.toLowerCase().includes(useComponentSearchStore().searchQuery.toLowerCase())
+    ); 
   }
 };
 
-return {cards,  matchesSearch};
+return {cards, matchesSearch};
 }
