+++ components/Cards/Stefan/Styled/Styled.vue
@@ -9,7 +9,7 @@
             <div style="text-align: left;">
               <v-row style="color: #656cbe;">Author: {{ cards[cardId].author }}
               </v-row>
-              <v-row style="color: #656cbe;">Co-Author: {{ cards[cardId].coAuthor }}</v-row>
+              <v-row style="color: #656cbe;">Co-Author/s: {{ cards[cardId].coAuthor.join(", ") }}</v-row>
             </div>
           </v-container>
         </v-col>
@@ -252,14 +252,14 @@
 </template>
 
 <script setup>
-import { useMyStyledCardStore } from '~/stores/Cards/StyledCard';
+import { useMyStyledCardStore } from '~/stores/Cards/Stefan/StyledCard';
 
 //Component Variables
 const useCardStore = useMyStyledCardStore();
 const cardId = 0; //Search card ID
 
 //Search Button Logic
-import { useSearchCards } from '~/components/Cards/SearchCards';
+import { useSearchCards } from '~/components/Cards/Stefan/StefanSearchCards';
 const { cards, matchesSearch } = useSearchCards();
