+++ components/Selects/Stefan/StefanSelectsPage.vue
@@ -1,0 +1,11 @@
+<template>
+    <StyledSelect />
+    <ActivitySelect />
+</template>
+
+<script setup>
+import StyledSelect from '~/components/Selects/Stefan/Styled/Styled.vue';
+import ActivitySelect from '~/components/Selects/Stefan/Activity/Activity.vue';
+</script>
+
+<style lang="scss" scoped></style>
