+++ components/Selects/Stefan/Styled/Styled.vue
@@ -10,7 +10,7 @@
             <div style="text-align: left;">
               <v-row style="color: #656cbe;">Author: {{ cards[cardId].author }}
               </v-row>
-              <v-row style="color: #656cbe;">Co-Author: {{ cards[cardId].coAuthor }}</v-row>
+              <v-row style="color: #656cbe;">Co-Author/s: {{ cards[cardId].coAuthor.join(", ") }}</v-row>
             </div>
           </v-container>
         </v-col>
@@ -169,14 +169,14 @@
 </template>
 
 <script setup>
-import { useMySelectStyledStore } from '~/stores/Selects/selectStyled';
+import { useMySelectStyledStore } from '~/stores/Selects/Stefan/selectStyled';
 
 //Component Variables
 const useSelectsStore = useMySelectStyledStore();
 const cardId = 0; //Search card ID
 
 //Search Button Logic
-import { useSearchSelects } from '~/components/Selects/SearchSelects';
+import { useSearchSelects } from '~/components/Selects/Stefan/StefanSearchSelects';
 const { cards, matchesSearch } = useSearchSelects();
