+++ stores/Cards/Stefan/InteractiveCard.ts
