+++ stores/Cards/Stefan/SlotsCard.ts
