+++ stores/Cards/Stefan/StyledCard.ts
