+++ stores/Selects/Stefan/selectActivity.ts
